INS-2216: remove packet wrapper and second future
